documenters: call to_schema in import_object
Sometimes, for unclear reasons, the call to `to_schema` fails at the
field documenter. The error is ultimately that the `Type` type does not
exists in the typing package which makes litle sense.

After investigating, it turns out that autodoc patches the typing
library at some point during the generate process. This conflicts with
the pandera model logic which relies on type annotations to define
column types.

To make sure the rendering process works properly, we call the `to_schema`
method early on before the typing patch happens.

This is a hack so it's probably fragile.